feat: Add comments (#61)
Add comments to posts, including the ability to subscribe and unsubscribe to notifications.

By default notifications are turned on for the author and no other users. Users can subscribe to them via the action in the comments widget.

Comments also contain draft functionality similar to posts, and work very similarly. In the future we'd ideally limit what markdown can be used (we dont need big headers, for example).

This also corrects an issue w/ the mail test harness (race condition).